Test_normalize_content moved to unit tests
