+++ test/integration/ggrc/models/test_revision.py
@@ -590,37 +590,6 @@ class TestRevisions(query_helper.WithQueryApi, TestCase):
     self.assertFalse(revisions[0].is_empty)
     self.assertTrue(revisions[1].is_empty)
 
-  @ddt.data({
-      'id': 2676L, 'object_type': u'Assessment', 'object_id': 19L,
-      'base_id': 2676L, 'created_at': datetime(2019, 10, 17, 9, 44, 5),
-      'updated_at': datetime(2019, 10, 17, 9, 44, 5), 'modified_by_id': None,
-      'display_name': '', 'type': 'AccessControlList', 'context_id': None,
-      'ac_role_id': 73L, 'parent_id': None, 'parent_id_nn': 0L,
-      'modified_by': None
-  },
-      {
-      u'parent_id': None, u'modified_by_id': None, u'object_id': 19,
-      u'created_at': u'2019-10-17T09:44:05', u'base_id': 2676,
-      u'updated_at': u'2019-10-17T09:44:05', u'object_type': u'Assessment',
-      u'display_name': u'', u'type': u'AccessControlList', u'context_id': None,
-      u'ac_role_id': 73, u'id': 2676, u'parent_id_nn': 0, u'modified_by': None
-  })
-  @ddt.unpack
-  def test_normalize_content(self, **content):
-    """Test converting new revision data"""
-    expected_result = {u'parent_id': None, u'modified_by_id': None,
-                       u'object_id': 19, u'created_at': u'2019-10-17T09:44:05',
-                       u'base_id': 2676, u'updated_at': u'2019-10-17T09:44:05',
-                       u'object_type': u'Assessment', u'display_name': u'',
-                       u'type': u'AccessControlList', u'context_id': None,
-                       u'ac_role_id': 73, u'id': 2676, u'parent_id_nn': 0,
-                       u'modified_by': None}
-
-    # pylint: disable=protected-access
-    result = ggrc.utils.revisions_diff.builder._normalize_content(content)
-
-    self.assertEqual(result, expected_result)
-
   def test_changes_revision(self):
     """Test revision is marked empty or not depending on changes"""
     with factories.single_commit():
